feat: criação de MaquinaController, metodos get e post

+++ src/main/java/br/com/matheus/manutencao/repository/MaquinaRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface MaquinaRepository extends JpaRepository <Maquina, Long> {
 
-    Optional<Maquina> findByNp(Integer np);
+    Optional<Maquina> findByNp(Long np);
 
 }
